refactor: Cars 일급 컬렉션 Class 생성
Cars 생성 -> Winner 객체가 필요없다고 판단해서 삭제
